pingenator: use -c1 in addition to -w1
I've historically used -w1 to check reachability in a script, but I have
seen 2 pings come back once in a while on a system under heavy load.  I
assume this is because the 1 second timeout gets scheduled late, on the
order of 1.25 seconds, and in that extra .25 seconds a 2nd ping is sent and
received.

I normally wouldn't care... but if we're doing all kinds of calculations
based on the assumption that each process sends exactly 1 ping, we should
probably make sure that's the reality.